click plumbing qr code/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
qrcode==7.4.2
Pillow>=10.2.0
python-dateutil==2.8.2
Flask-Admin==1.6.1
Flask-Security-Too==5.1.2
bcrypt==4.0.1
email-validator==2.0.0.post2
